Share sample code between Framework and .NET 5 sample apps
