+++ BatchProcessor.sln
@@ -20,7 +20,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{687DB7
 		SharedAssemblyInfo.cs = SharedAssemblyInfo.cs
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestAppNet5", "TestAppNet5\TestAppNet5.csproj", "{1D1E8C34-19DA-47D5-8918-C5E5B96512EF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestAppNet5", "TestAppNet5\TestAppNet5.csproj", "{1D1E8C34-19DA-47D5-8918-C5E5B96512EF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestShared", "TestShared\TestShared.csproj", "{450B53AE-F014-4EF7-96F5-EE6343F73F5B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,6 +58,10 @@ Global
 		{1D1E8C34-19DA-47D5-8918-C5E5B96512EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D1E8C34-19DA-47D5-8918-C5E5B96512EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1D1E8C34-19DA-47D5-8918-C5E5B96512EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{450B53AE-F014-4EF7-96F5-EE6343F73F5B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{450B53AE-F014-4EF7-96F5-EE6343F73F5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{450B53AE-F014-4EF7-96F5-EE6343F73F5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{450B53AE-F014-4EF7-96F5-EE6343F73F5B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
